Dashboard/Backend_deploy/pyproject.toml
[tool.poetry]
name = "fastapi-dynamodb"
version = "0.1.0"
description = ""
authors = ["Piotr Rogulski <rivinek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.72.0"
pynamodb = "^6.0.1"
uvicorn = "^0.17.0"
minio = "^7.2.7"
datetime = "^5.5"
torch = "^2.4.0"
torchvision = "^0.19.0"
matplotlib = "^3.9.1"
tqdm = "^4.66.5"
pillow-heif = "^0.18.0"
python-multipart = "^0.0.9"
pandas = "^2.2.2"
boto3 = "^1.34.159"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.12b0"
isort = "^5.10.1"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"